backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
opencv-python
torch
torchvision
regex